Initial commit of beamline-specific code
